Added unit tests for dice flag conversion and removed full text index code
The full text index doesn't get properly created by `engine.create_all()` which means that it breaks the tests. Since it's ignored by Alembic, anyway, there's little reason to keep it in the codebase, as a result (plus it will probably take longer to generate the index than to just query against the non-indexed values in filtration tests).