Removed `.env` support for production
This isn't actually working at all locally, either! Turns out Compose is injecting the .env variables into the environment, and Pydantic is picking that up (because it doesn't have the necessary module for loading .env files on its own at the moment). Could potentially change this, but if I do I think I want to load the path from an environment variable rather than hard-coding it.