Updated handling of related cards for card details output
In the legacy behavior, I wasn't getting consistent output for that list because it changed what was in the output based on whether we were looking at a Phoenixborn, Phoenixborn card, etc.

Now, it outputs different lists depending on what we're looking at, which means that for everything except Ancestor Spirit we get completely consistent behavior (Ancestor Spirit is a weird one, because it's summoned by Rimea's unique, but *isn't* itself unique).