Changed the order in which the server is toggled off and then on
The crontab command that's running could easily restart the server while it's updating things, so there's not much point in trying to ensure it's completely offline, and the dependency download and migrations should generally not require it to be down (since most everything that matters will just be in memory). Instead, we'll simply toggle the server after all upgrade tasks are complete.